Extract mod to its own file
