+++ sway-parse/src/lib.rs
@@ -13,6 +13,8 @@ mod pattern;
 mod priv_prelude;
 mod punctuated;
 mod submodule;
+#[cfg(test)]
+mod test_utils;
 mod token;
 mod ty;
 mod where_clause;
@@ -58,33 +60,4 @@ pub fn parse_module_kind(
         parser.parse::<DocComment>()?;
     }
     parser.parse()
-}
-
-#[cfg(test)]
-mod test_utils {
-    use crate::{priv_prelude::ParseToEnd, Parse, Parser};
-    use std::sync::Arc;
-
-    pub fn parse<T>(input: &str) -> T
-    where
-        T: Parse,
-    {
-        let handler = <_>::default();
-        let ts = crate::token::lex(&handler, &Arc::from(input), 0, input.len(), None).unwrap();
-        Parser::new(&handler, &ts)
-            .parse()
-            .unwrap_or_else(|_| panic!("Parse error: {:?}", handler.consume().0))
-    }
-
-    pub fn parse_to_end<T>(input: &str) -> T
-    where
-        T: ParseToEnd,
-    {
-        let handler = <_>::default();
-        let ts = crate::token::lex(&handler, &Arc::from(input), 0, input.len(), None).unwrap();
-        Parser::new(&handler, &ts)
-            .parse_to_end()
-            .map(|(m, _)| m)
-            .unwrap_or_else(|_| panic!("Parse error: {:?}", handler.consume().0))
-    }
 }
